+++ src/constants.js
@@ -1,0 +1,3 @@
+import axios from "axios";
+
+export const DEFAULT_AXIOS = axios.create();