+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.0.0 (In Development)
 
+### 1.0.0-alpha.2 (2020-03-23)
+
+- Provide ability to override `SadnessProvider` default context values with
+  `SadnessConfig` component
+
 ### 1.0.0-alpha.1 (2020-03-23)
 
 - Use `package.json#source` field for pointing on source data, not